--- src/interceptors/authentication.interceptor.ts
@@ -1,525 +1,0 @@
-import {
-  injectable,
-  Interceptor,
-  InvocationContext,
-  InvocationResult,
-  Provider,
-  service,
-  ValueOrPromise,
-} from '@loopback/core';
-import {AnyObject, repository} from '@loopback/repository';
-import {HttpErrors} from '@loopback/rest';
-import {
-  ActivityLogType,
-  MethodType,
-  PermissionKeys,
-  ReferenceType,
-} from '../enums';
-import {
-  Credential,
-  RequestCreateNewUser,
-  RequestCreateNewUserByEmail,
-  RequestCreateNewUserByWallet,
-  RequestLoginByOTP,
-  User,
-  UserWithRelations,
-  Wallet,
-} from '../models';
-import {
-  ActivityLogRepository,
-  NetworkRepository,
-  RequestCreateNewUserByEmailRepository,
-  UserOTPRepository,
-  UserRepository,
-  WalletRepository,
-} from '../repositories';
-import {
-  CurrencyService,
-  FriendService,
-  MetricService,
-  UserOTPService,
-} from '../services';
-import {securityId, UserProfile} from '@loopback/security';
-import {assign, intersection} from 'lodash';
-import NonceGenerator from 'a-nonce-generator';
-import {validateAccount} from '../utils/validate-account';
-import {isHex} from '@polkadot/util';
-import validator from 'validator';
-import {config} from '../config';
-import {generateObjectId} from '../utils/formatter';
-import {PolkadotJs} from '../utils/polkadot-js';
-
-/**
- * This class will be bound to the application as an `Interceptor` during
- * `boot`
- */
-@injectable({tags: {key: AuthenticationInterceptor.BINDING_KEY}})
-export class AuthenticationInterceptor implements Provider<Interceptor> {
-  static readonly BINDING_KEY = `interceptors.${AuthenticationInterceptor.name}`;
-
-  constructor(
-    @repository(ActivityLogRepository)
-    protected activityLogRepository: ActivityLogRepository,
-    @repository(RequestCreateNewUserByEmailRepository)
-    protected requestCreateNewUserByEmailRepository: RequestCreateNewUserByEmailRepository,
-    @repository(NetworkRepository)
-    private networkRepository: NetworkRepository,
-    @repository(UserRepository)
-    private userRepository: UserRepository,
-    @repository(UserOTPRepository)
-    private userOTPRepository: UserOTPRepository,
-    @repository(WalletRepository)
-    private walletRepository: WalletRepository,
-    @service(CurrencyService)
-    private currencyService: CurrencyService,
-    @service(FriendService)
-    private friendService: FriendService,
-    @service(MetricService)
-    private metricService: MetricService,
-    @service(UserOTPService)
-    private userOTPService: UserOTPService,
-  ) {}
-
-  /**
-   * This method is used by LoopBack context to produce an interceptor function
-   * for the binding.
-   *
-   * @returns An interceptor function
-   */
-  value() {
-    return this.intercept.bind(this);
-  }
-
-  /**
-   * The logic to intercept an invocation
-   * @param invocationCtx - Invocation context
-   * @param next - A function to invoke next interceptor or the target method
-   */
-  async intercept(
-    invocationCtx: InvocationContext,
-    next: () => ValueOrPromise<InvocationResult>,
-  ) {
-    await this.beforeAuthenticate(invocationCtx);
-
-    const result = await next();
-
-    await this.afterAuthenticate(invocationCtx, result);
-
-    return result;
-  }
-
-  async beforeAuthenticate(invocationCtx: InvocationContext): Promise<void> {
-    const methodName = invocationCtx.methodName as MethodType;
-
-    if (methodName.startsWith('signup')) {
-      const {username} = invocationCtx.args[0] as RequestCreateNewUser;
-
-      this.validateUsername(username);
-
-      const or: AnyObject[] = [{username}];
-
-      if (invocationCtx.args[0]?.email) {
-        or.push({email: invocationCtx.args[0].email});
-      }
-
-      const foundUser = await this.userRepository.find({
-        where: {or},
-      });
-
-      if (foundUser.length > 0) {
-        throw new HttpErrors.UnprocessableEntity(
-          'Username/EmailAlreadyRegistered',
-        );
-      }
-    }
-
-    let user: User | UserWithRelations | undefined | null;
-
-    switch (methodName) {
-      // Sign up with wallet
-      case MethodType.SIGNUPBYWALLET: {
-        const {address, network, name} = invocationCtx
-          .args[0] as RequestCreateNewUserByWallet;
-        const fixedAddress = isHex(`0x${address}`) ? `0x${address}` : address;
-
-        await this.validateWalletAddress(network, fixedAddress);
-
-        const existingWallet = await this.walletRepository.exists(fixedAddress);
-
-        if (existingWallet) {
-          throw new HttpErrors.UnprocessableEntity(
-            'Wallet address already exists',
-          );
-        }
-
-        const existingNetwork = await this.networkRepository.exists(network);
-
-        if (!existingNetwork) {
-          throw new HttpErrors.UnprocessableEntity('Network not exists');
-        }
-
-        Object.assign(invocationCtx.args[0], {
-          id: generateObjectId(),
-          name: name.substring(0, 22),
-          permissions: this.getPermissions(fixedAddress),
-          fullAccess: true,
-        });
-
-        invocationCtx.args[1] = new Wallet({
-          id: fixedAddress,
-          networkId: network,
-          primary: true,
-        });
-
-        break;
-      }
-
-      // Signup with email
-      case MethodType.SIGNUPBYEMAIL: {
-        const {name, email} = invocationCtx
-          .args[0] as RequestCreateNewUserByEmail;
-
-        if (!validator.isEmail(email)) {
-          throw new HttpErrors.UnprocessableEntity('Invalid email address');
-        }
-
-        Object.assign(invocationCtx.args[0], {
-          id: generateObjectId(),
-          name: name.substring(0, 22),
-        });
-
-        break;
-      }
-
-      // Login with wallet
-      case MethodType.LOGINBYWALLET: {
-        const credential = invocationCtx.args[0] as Credential;
-        const {nonce, networkType} = credential;
-        const [publicAddress, account] = credential.publicAddress.split('/');
-        const nearAccount = isHex(`0x${account}`) ? `0x${account}` : account;
-
-        if (nonce === 0 || !nonce) {
-          throw new HttpErrors.Unauthorized('Invalid nonce!');
-        }
-
-        if (!credential?.role) credential.role = 'user';
-
-        const currentNetwork = await this.networkRepository.findById(
-          networkType,
-        );
-        const networks = await this.networkRepository.find({
-          where: {blockchainPlatform: currentNetwork.blockchainPlatform},
-        });
-        const networkIds = networks.map(network => network.id);
-        const wallet = await this.walletRepository.findOne({
-          where: {
-            id: nearAccount ?? publicAddress,
-            networkId: {inq: networkIds},
-          },
-          include: ['user'],
-        });
-
-        if (!wallet?.user) {
-          throw new HttpErrors.Unauthorized('Wallet address not exists!');
-        }
-
-        user = wallet.user;
-
-        if (user.nonce !== nonce) {
-          throw new HttpErrors.Unauthorized('Invalid nonce!');
-        }
-
-        const verified = await validateAccount(
-          assign(credential, {publicAddress}),
-          currentNetwork,
-          wallet.id,
-        );
-
-        if (!verified) {
-          throw new HttpErrors.Unauthorized('[auth] Failed to verified!');
-        }
-
-        invocationCtx.args[1] = wallet.id;
-
-        break;
-      }
-
-      // Login with OTP
-      case MethodType.LOGINBYOTP: {
-        const {token} = invocationCtx.args[0] as RequestLoginByOTP;
-
-        const validOTP = await this.userOTPService.verifyOTP(token);
-
-        if (!validOTP) {
-          throw new HttpErrors.Unauthorized('OTP invalid or expired!');
-        }
-
-        const key = `sign-up/${token}`;
-        const newUser = await this.requestCreateNewUserByEmailRepository.get(
-          key,
-        );
-
-        if (newUser?.id === validOTP.userId.toString()) {
-          const users = await this.userRepository.find({
-            where: {
-              or: [{email: newUser.email}, {username: newUser.username}],
-            },
-          });
-
-          if (users.length > 0) {
-            throw new HttpErrors.UnprocessableEntity('UserAlreadyExists');
-          }
-
-          user = await this.userRepository.create(newUser);
-          invocationCtx.args[2] = user;
-        } else {
-          user = await this.userRepository.findOne({
-            where: {
-              id: validOTP.userId,
-            },
-          });
-        }
-
-        invocationCtx.args[1] = validOTP.userId;
-
-        break;
-      }
-    }
-
-    if (methodName.startsWith('login')) {
-      if (!user) throw new HttpErrors.UnprocessableEntity('User not exists!');
-      if (methodName === MethodType.LOGINBYWALLET) {
-        const credential = invocationCtx.args[0] as Credential;
-
-        if (credential.role === 'admin') {
-          const [permission] = intersection(user.permissions, [
-            PermissionKeys.ADMIN,
-          ]);
-
-          if (permission !== PermissionKeys.ADMIN) {
-            throw new HttpErrors.Forbidden('Invalid admin');
-          }
-        } else {
-          const [userPermission] = intersection(user.permissions, [
-            PermissionKeys.USER,
-          ]);
-
-          if (userPermission !== PermissionKeys.USER) {
-            throw new HttpErrors.Forbidden('Invalid user');
-          }
-        }
-      }
-
-      const userProfile: UserProfile = {
-        [securityId]: user.id!.toString(),
-        id: user.id,
-        name: user.name,
-        username: user.username,
-        createdAt: user.createdAt,
-        permissions: user.permissions,
-      };
-
-      invocationCtx.args[0].data = userProfile;
-    }
-  }
-
-  async afterAuthenticate(
-    invocationCtx: InvocationContext,
-    result: AnyObject,
-  ): Promise<void> {
-    const methodName = invocationCtx.methodName as MethodType;
-    const jobs = [];
-
-    switch (methodName) {
-      // Sign up with wallet
-      case MethodType.SIGNUPBYWALLET: {
-        const wallet = invocationCtx.args[1] as Wallet;
-        jobs.push(
-          this.userRepository.accountSetting(result.id).create({}),
-          this.userRepository.notificationSetting(result.id).create({}),
-          this.userRepository.languageSetting(result.id).create({}),
-          this.metricService.countServerMetric(),
-          this.activityLogRepository.create({
-            type: ActivityLogType.NEWUSER,
-            userId: result.id,
-            referenceId: result.id,
-            referenceType: ReferenceType.USER,
-          }),
-          this.userRepository.wallets(result.id).create(wallet),
-          this.currencyService.create(result.id, wallet.networkId),
-        );
-        break;
-      }
-
-      // Login by wallet
-      case MethodType.LOGINBYWALLET: {
-        // Generate random nonce after login
-        const [credential, walletId] = invocationCtx.args;
-        const networkId = credential.networkType;
-        const userId = credential.data.id;
-        const ng = new NonceGenerator();
-        const newNonce = ng.generate();
-
-        jobs.push(
-          this.userRepository.updateById(userId, {
-            nonce: newNonce,
-            fullAccess: true,
-          }),
-        );
-
-        if (credential.role === 'user') {
-          await this.walletRepository.updateAll({primary: false}, {userId});
-          await this.currencyService.update(userId, networkId);
-          jobs.push(
-            this.walletRepository.updateById(walletId, {
-              primary: true,
-              networkId,
-            }),
-          );
-        }
-
-        break;
-      }
-
-      // Login by email
-      case MethodType.LOGINBYOTP: {
-        const {token} = invocationCtx.args[0] as RequestLoginByOTP;
-        const userId = invocationCtx.args[1];
-        const key = `sign-up/${token}`;
-
-        if (invocationCtx.args[2]) {
-          const newUser = invocationCtx.args[2] as User;
-          jobs.push(
-            this.userRepository.accountSetting(newUser.id).create({}),
-            this.userRepository.notificationSetting(newUser.id).create({}),
-            this.userRepository.languageSetting(newUser.id).create({}),
-            this.metricService.countServerMetric(),
-            this.activityLogRepository.create({
-              type: ActivityLogType.NEWUSER,
-              userId: newUser.id,
-              referenceId: newUser.id,
-              referenceType: ReferenceType.USER,
-            }),
-          );
-        }
-
-        await this.walletRepository.updateAll(
-          {networkId: 'myriad', primary: true},
-          {userId},
-        );
-        jobs.push(
-          this.requestCreateNewUserByEmailRepository.delete(key),
-          this.userOTPRepository.deleteAll({
-            token: token,
-          }),
-        );
-        break;
-      }
-    }
-
-    Promise.allSettled(jobs) as Promise<AnyObject>;
-  }
-
-  validateUsername(username: string): void {
-    if (username[username.length - 1] === '_') {
-      throw new HttpErrors.UnprocessableEntity(
-        'Last character must be an ascii letter (a-z) or number (0-9)',
-      );
-    }
-
-    if (username[0] === '_') {
-      throw new HttpErrors.UnprocessableEntity(
-        'Character must be start from an ascii letter (a-z) or number (0-9)',
-      );
-    }
-
-    if (!username.match('^[a-z0-9_]+$')) {
-      throw new HttpErrors.UnprocessableEntity(
-        'Only allowed ascii letter (a-z), number (0-9), and underscore(_)',
-      );
-    }
-  }
-
-  async validateWalletAddress(network: string, id: string): Promise<void> {
-    switch (network) {
-      case 'near':
-        return this.validateNearAddress(id);
-
-      case 'ethereum': {
-        if (id.length !== 42 && !isHex(id)) {
-          throw new HttpErrors.UnprocessableEntity('Invalid ethereum address');
-        }
-
-        return;
-      }
-      default: {
-        const polkadotJs = new PolkadotJs();
-        const valid = polkadotJs.validatePolkadotAddress(id);
-
-        if (id.length !== 66 && !valid) {
-          throw new HttpErrors.UnprocessableEntity('Invalid polkadot address');
-        }
-
-        return;
-      }
-    }
-  }
-
-  async validateNearAddress(id: string): Promise<void> {
-    if (isHex(id)) return;
-
-    const nearNetwork = await this.networkRepository.findById('near');
-    const environment = nearNetwork.rpcURL.split('.')[1];
-
-    let nearId = '';
-    let nearStatus = false;
-
-    switch (environment) {
-      case 'betanet':
-        nearStatus = id.endsWith('.betanet');
-        nearId = id.split('.betanet')[0];
-        break;
-
-      case 'testnet':
-        nearStatus = id.endsWith('.testnet');
-        nearId = id.split('.testnet')[0];
-        break;
-
-      default:
-        nearStatus = id.endsWith('.near');
-        nearId = id.split('.near')[0];
-        break;
-    }
-
-    if (!nearStatus) {
-      throw new HttpErrors.UnprocessableEntity('Invalid near id');
-    }
-
-    if (!nearId.match('^[a-z0-9_-]+$')) {
-      throw new HttpErrors.UnprocessableEntity(
-        'Only allowed ascii letter (a-z), number (0-9), dash(-) and underscore(_)',
-      );
-    }
-
-    return;
-  }
-
-  getPermissions(registeredAddress: string): PermissionKeys[] {
-    const permissions: PermissionKeys[] = [PermissionKeys.USER];
-
-    try {
-      const {getKeyring, getHexPublicKey} = new PolkadotJs();
-      const mnemonic = config.MYRIAD_ADMIN_SUBSTRATE_MNEMONIC;
-      const serverAdmin = getKeyring().addFromMnemonic(mnemonic);
-      const address = getHexPublicKey(serverAdmin);
-
-      if (registeredAddress === address) {
-        permissions.push(PermissionKeys.MASTER, PermissionKeys.ADMIN);
-      }
-    } catch {
-      // ignore
-    }
-
-    return permissions;
-  }
-}
