+++ src/interceptors/index.ts
@@ -1,4 +1,3 @@
-export * from './authentication.interceptor';
 export * from './authorize.interceptor';
 export * from './find-by-id.interceptor';
 export * from './pagination.interceptor';
